Describe use of actions-riff-raff in readme
When deploying a static site recently for
[datawrapper-import](https://github.com/guardian/datawrapper-import), it
wasn’t clear to me that the repository needed to have the necessary
access for Riff Raff to successfully write files for the
`deploy:datawrapper-import` project. Hopefully this message makes it
clearer that that access needs to be set up when setting up a new static
site.